[Refactor] #94 LIVD-57 - 날짜 범위 포맷팅 로직 공통화
